fix: Hardcode chalkan3 user in dotfiles.sls
The dotfiles.sls state was using pillar data for the user and home
directory when cloning the repository and applying stow. This could
lead to incorrect configurations if the install.sh script was run
by a user other than 'chalkan3'.

This commit hardcodes 'chalkan3' as the target user and '/home/chalkan3'
as the target directory for the git clone and stow operations,
ensuring the dotfiles are always configured for the intended user.